+++ packages/flp-config-inquirer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/flp-config-inquirer
 
+## 0.2.69
+
+### Patch Changes
+
+-   Updated dependencies [c81864a]
+    -   @sap-ux/adp-tooling@0.13.24
+
+## 0.2.68
+
+### Patch Changes
+
+-   Updated dependencies [4f0eaae]
+    -   @sap-ux/adp-tooling@0.13.23
+
 ## 0.2.67
 
 ### Patch Changes